'Refactored by Sourcery'

+++ Cutiepii_Robot/modules/anime.py
@@ -23,7 +23,7 @@ close_btn = "❌"
 def shorten(description, info='anilist.co'):
     msg = ""
     if len(description) > 700:
-        description = description[0:500] + '....'
+        description = f'{description[:500]}....'
         msg += f"\n*Description*: _{description}_[Read More]({info})"
     else:
         msg += f"\n*Description*:_{description}_"
@@ -34,15 +34,17 @@ def shorten(description, info='anilist.co'):
 def t(milliseconds: int) -> str:
     """Inputs time in milliseconds, to get beautified time,
     as string"""
-    seconds, milliseconds = divmod(int(milliseconds), 1000)
+    seconds, milliseconds = divmod(milliseconds, 1000)
     minutes, seconds = divmod(seconds, 60)
     hours, minutes = divmod(minutes, 60)
     days, hours = divmod(hours, 24)
-    tmp = ((str(days) + " Days, ") if days else "") + \
-        ((str(hours) + " Hours, ") if hours else "") + \
-        ((str(minutes) + " Minutes, ") if minutes else "") + \
-        ((str(seconds) + " Seconds, ") if seconds else "") + \
-        ((str(milliseconds) + " ms, ") if milliseconds else "")
+    tmp = (
+        (f"{str(days)} Days, " if days else "")
+        + (f"{str(hours)} Hours, " if hours else "")
+        + (f"{str(minutes)} Minutes, " if minutes else "")
+        + (f"{str(seconds)} Seconds, " if seconds else "")
+        + (f"{str(milliseconds)} ms, " if milliseconds else "")
+    )
     return tmp[:-2]
 
 
@@ -220,7 +222,7 @@ def anime(update, context):
             trailer_id = trailer.get('id', None)
             site = trailer.get('site', None)
             if site == "youtube":
-                trailer = 'https://youtu.be/' + trailer_id
+                trailer = f'https://youtu.be/{trailer_id}'
         description = json.get('description', 'N/A').replace('<i>', '').replace(
             '</i>', '').replace('<br>', '')
         msg += shorten(description, info)
@@ -278,8 +280,7 @@ def character(update, context):
         site_url = json.get('siteUrl')
         char_name = f"{json.get('name').get('full')}"
         msg += shorten(description, site_url)
-        image = json.get('image', None)
-        if image:
+        if image := json.get('image', None):
             image = image.get('large')
             buttons = [[InlineKeyboardButton("Add To Favorite Character", callback_data=f"xanime_fvrtchar={char_name}")]]
             update.effective_message.reply_photo(
@@ -499,18 +500,14 @@ def change_quote(update: Update, context: CallbackContext):
 
 @run_async
 def watchlist(update, context):
-    chat = update.effective_chat  
-    user = update.effective_user 
-    message = update.effective_message  
-    watchlist = list(REDIS.sunion(f'anime_watch_list{user.id}'))
-    watchlist.sort()
-    watchlist = "\n• ".join(watchlist)
-    if watchlist:
+    chat = update.effective_chat
+    user = update.effective_user
+    message = update.effective_message
+    watchlist = sorted(REDIS.sunion(f'anime_watch_list{user.id}'))
+    if watchlist := "\n• ".join(watchlist):
         message.reply_text(
-            "{}<b>'s Watchlist:</b>"
-            "\n• {}".format(mention_html(user.id, user.first_name),
-                        watchlist),
-            parse_mode=ParseMode.HTML
+            f"{mention_html(user.id, user.first_name)}<b>'s Watchlist:</b>\n• {watchlist}",
+            parse_mode=ParseMode.HTML,
         )
     else:
         message.reply_text(
@@ -543,18 +540,14 @@ def removewatchlist(update, context):
 
 @run_async
 def fvrtchar(update, context):
-    chat = update.effective_chat  
-    user = update.effective_user 
-    message = update.effective_message  
-    fvrt_char = list(REDIS.sunion(f'anime_fvrtchar{user.id}'))
-    fvrt_char.sort()
-    fvrt_char = "\n• ".join(fvrt_char)
-    if fvrt_char:
+    chat = update.effective_chat
+    user = update.effective_user
+    message = update.effective_message
+    fvrt_char = sorted(REDIS.sunion(f'anime_fvrtchar{user.id}'))
+    if fvrt_char := "\n• ".join(fvrt_char):
         message.reply_text(
-            "{}<b>'s Favorite Characters List:</b>"
-            "\n• {}".format(mention_html(user.id, user.first_name),
-                        fvrt_char),
-            parse_mode=ParseMode.HTML
+            f"{mention_html(user.id, user.first_name)}<b>'s Favorite Characters List:</b>\n• {fvrt_char}",
+            parse_mode=ParseMode.HTML,
         )
     else:
         message.reply_text(
@@ -588,18 +581,14 @@ def removefvrtchar(update, context):
     
 @run_async
 def readmanga(update, context):
-    chat = update.effective_chat  
-    user = update.effective_user 
-    message = update.effective_message  
-    manga_list = list(REDIS.sunion(f'anime_mangaread{user.id}'))
-    manga_list.sort()
-    manga_list = "\n• ".join(manga_list)
-    if manga_list:
+    chat = update.effective_chat
+    user = update.effective_user
+    message = update.effective_message
+    manga_list = sorted(REDIS.sunion(f'anime_mangaread{user.id}'))
+    if manga_list := "\n• ".join(manga_list):
         message.reply_text(
-            "{}<b>'s Manga Lists:</b>"
-            "\n• {}".format(mention_html(user.id, user.first_name),
-                        manga_list),
-            parse_mode=ParseMode.HTML
+            f"{mention_html(user.id, user.first_name)}<b>'s Manga Lists:</b>\n• {manga_list}",
+            parse_mode=ParseMode.HTML,
         )
     else:
         message.reply_text(
@@ -636,22 +625,10 @@ def animestuffs(update, context):
     user = update.effective_user
     splitter = query.data.split('=')
     query_match = splitter[0]
-    callback_anime_data = splitter[1] 
-    if query_match == "xanime_watchlist":
-        watchlist = list(REDIS.sunion(f'anime_watch_list{user.id}'))
-        if not callback_anime_data in watchlist:
-            REDIS.sadd(f'anime_watch_list{user.id}', callback_anime_data)
-            context.bot.answer_callback_query(query.id,
-                                                text=f"{callback_anime_data} is successfully added to your watch list.",
-                                                show_alert=True)
-        else:
-            context.bot.answer_callback_query(query.id,
-                                                text=f"{callback_anime_data} is already exists in your watch list!",
-                                                show_alert=True)
-            
-    elif query_match == "xanime_fvrtchar":   
+    callback_anime_data = splitter[1]
+    if query_match == "xanime_fvrtchar":
         fvrt_char = list(REDIS.sunion(f'anime_fvrtchar{user.id}'))
-        if not callback_anime_data in fvrt_char:
+        if callback_anime_data not in fvrt_char:
             REDIS.sadd(f'anime_fvrtchar{user.id}', callback_anime_data)
             context.bot.answer_callback_query(query.id,
                                                 text=f"{callback_anime_data} is successfully added to your favorite character.",
@@ -660,9 +637,9 @@ def animestuffs(update, context):
             context.bot.answer_callback_query(query.id,
                                                 text=f"{callback_anime_data} is already exists in your favorite characters list!",
                                                 show_alert=True)
-    elif query_match == "xanime_manga":   
+    elif query_match == "xanime_manga":
         fvrt_char = list(REDIS.sunion(f'anime_mangaread{user.id}'))
-        if not callback_anime_data in fvrt_char:
+        if callback_anime_data not in fvrt_char:
             REDIS.sadd(f'anime_mangaread{user.id}', callback_anime_data)
             context.bot.answer_callback_query(query.id,
                                                 text=f"{callback_anime_data} is successfully added to your read list.",
@@ -670,6 +647,18 @@ def animestuffs(update, context):
         else:
             context.bot.answer_callback_query(query.id,
                                                 text=f"{callback_anime_data} is already exists in your read list!",
+                                                show_alert=True)
+
+    elif query_match == "xanime_watchlist":
+        watchlist = list(REDIS.sunion(f'anime_watch_list{user.id}'))
+        if callback_anime_data not in watchlist:
+            REDIS.sadd(f'anime_watch_list{user.id}', callback_anime_data)
+            context.bot.answer_callback_query(query.id,
+                                                text=f"{callback_anime_data} is successfully added to your watch list.",
+                                                show_alert=True)
+        else:
+            context.bot.answer_callback_query(query.id,
+                                                text=f"{callback_anime_data} is already exists in your watch list!",
                                                 show_alert=True)
             
 
@@ -686,29 +675,32 @@ def button(update, context):
     user_and_admin_list = [original_user_id, OWNER_ID] + DRAGONS + DEV_USERS 
 
     bot.answer_callback_query(query.id)
-    if query_type == "anime_close":
-        if query.from_user.id in user_and_admin_list:
-            message.delete()
-        else:
-            query.answer("You are not allowed to use this.")
+    if (
+        query_type == "anime_close"
+        and query.from_user.id in user_and_admin_list
+    ):
+        message.delete()
+    elif (
+        query_type == "anime_close"
+        or query_type in ('anime_anime', 'anime_manga')
+        and query.from_user.id != original_user_id
+    ):
+        query.answer("You are not allowed to use this.")
     elif query_type in ('anime_anime', 'anime_manga'):
+        message.delete()
+        progress_message = bot.sendMessage(message.chat.id,
+                                           "Searching.... ")
         mal_id = data[2]
-        if query.from_user.id == original_user_id:
-            message.delete()
-            progress_message = bot.sendMessage(message.chat.id,
-                                               "Searching.... ")
-            caption, buttons, image = get_anime_manga(mal_id, query_type,
-                                                      original_user_id)
-            bot.sendPhoto(
-                message.chat.id,
-                photo=image,
-                caption=caption,
-                parse_mode=ParseMode.HTML,
-                reply_markup=InlineKeyboardMarkup(buttons),
-                disable_web_page_preview=False)
-            progress_message.delete()
-        else:
-            query.answer("You are not allowed to use this.")
+        caption, buttons, image = get_anime_manga(mal_id, query_type,
+                                                  original_user_id)
+        bot.sendPhoto(
+            message.chat.id,
+            photo=image,
+            caption=caption,
+            parse_mode=ParseMode.HTML,
+            reply_markup=InlineKeyboardMarkup(buttons),
+            disable_web_page_preview=False)
+        progress_message.delete()
